deprecate DP support, PTL support through 2bd54e46, PT support through 20230220, fix loop synchronization during resume

+++ src/fts_examples/ipynb_src/fts_superglue_nb.py
@@ -152,7 +152,6 @@
 # - ``ddp`` (and alias ``ddp_find_unused_parameters_false``)
 # - ``fsdp`` (and alias ``fsdp_cpu_offload``)
 # - ``ddp_spawn`` (and aliases ``ddp_fork``, ``ddp_notebook``)
-# - ``dp``
 #
 # Custom or officially unsupported strategies can be used by setting [FinetuningScheduler.allow_untested](https://finetuning-scheduler.readthedocs.io/en/stable/api/finetuning_scheduler.fts.html?highlight=allow_untested#finetuning_scheduler.fts.FinetuningScheduler.params.allow_untested) to ``True``.
 # Note that most currently unsupported strategies are so because they require varying degrees of modification to be compatible. For example, ``deepspeed`` will require a ``StrategyAdapter`` to be written (similar to the one for ``FSDP``, ``FSDPStrategyAdapter``) before support can be added (PRs welcome!),
@@ -581,7 +580,7 @@ def train() -> None:
         enable_progress_bar=enable_progress_bar,
         max_epochs=1,
         limit_train_batches=2,
-        precision=16,
+        precision="16-mixed",
         accelerator="auto",
         devices=1 if is_cuda_available() else None,
         callbacks=callbacks,
